requirements.txt
Flask==2.2.3
Flask-Migrate==3.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
boto3==1.35.69
botocore==1.35.69
gunicorn==23.0.0
task_priority_summary_verify_utils==0.1
